wip on tests

+++ test/default.nix
@@ -11,17 +11,21 @@ let
 
   # we need some output so circle doesn't break
   # print a line every minute
-  for i in $(seq 60); do
-    echo "tick still testing ($i)"
-    sleep 60
-  done &
-  _jid=$!
+  if [ ! -z ''${CIRCLECI+x} ]; then
+   for i in $(seq 60); do
+     echo "tick still testing ($i)"
+     sleep 60
+   done &
+   _jid=$!
+  fi;
 
   # alas, we cannot specify --features in the virtual workspace
   cargo test --manifest-path=crates/holochain/Cargo.toml --features slow_tests -- --nocapture
 
   # stop our background ticker
-  kill $_jid
+  if [ ! -z ''${CIRCLECI+x} ]]; then
+   kill $_jid
+  fi
   '';
 
   t-merge = pkgs.writeShellScriptBin "hc-merge-test"
